+++ cloud-hypervisor/src/main.rs
@@ -3,6 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
+mod logger;
 #[cfg(test)]
 mod test_util;
 
@@ -38,6 +39,8 @@ use vmm::vm_config::{
 };
 use vmm_sys_util::eventfd::EventFd;
 use vmm_sys_util::signal::block_signal;
+
+use crate::logger::Logger;
 
 #[cfg(feature = "dhat-heap")]
 #[global_allocator]
@@ -114,47 +117,6 @@ enum FdTableError {
     GetFd(#[source] std::io::Error),
     #[error("Failed to duplicate file handle")]
     Dup2(#[source] std::io::Error),
-}
-
-struct Logger {
-    output: Mutex<Box<dyn std::io::Write + Send>>,
-    start: std::time::Instant,
-}
-
-impl log::Log for Logger {
-    fn enabled(&self, _metadata: &log::Metadata) -> bool {
-        true
-    }
-
-    fn log(&self, record: &log::Record) {
-        if !self.enabled(record.metadata()) {
-            return;
-        }
-
-        let now = std::time::Instant::now();
-        let duration = now.duration_since(self.start);
-        let duration_s = duration.as_secs_f32();
-
-        let location = if let (Some(file), Some(line)) = (record.file(), record.line()) {
-            format!("{file}:{line}")
-        } else {
-            record.target().to_string()
-        };
-
-        let mut out = self.output.lock().unwrap();
-        write!(
-            &mut *out,
-            // 10: 6 decimal places + sep => whole seconds in range `0..=999` properly aligned
-            "cloud-hypervisor: {:>10.6?}s: <{}> {}:{} -- {}\r\n",
-            duration_s,
-            std::thread::current().name().unwrap_or("anonymous"),
-            record.level(),
-            location,
-            record.args(),
-        )
-        .ok();
-    }
-    fn flush(&self) {}
 }
 
 fn prepare_default_values() -> (String, String, String) {
